+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -96,7 +96,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-02-25T16:39:36.903Z")
+    created: Optional[datetime] = parse_datetime("2026-02-25T17:35:44.190Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")