+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -259,7 +259,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KJATQHYHRBX08N5J7P4EVHBR"
+    id: Optional[str] = "01KJAXYD6XXR5YGKXTFVPMJGRK"
 
     description: Optional[str] = None
 
@@ -1214,7 +1214,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KJATQHY8WE9YYQKSVZRRGM5S"
+        "tool_01KJAXYD6M50J8VRVEZ3YT78WZ"
     )
 
     display_name: Optional[str] = None
@@ -1345,7 +1345,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KJATQHY6Y1R0W6FR4B9HCJY7"
+    id: Optional[str] = "01KJAXYD6KRP732YMS6HKHJ7XE"
 
     description: Optional[str] = None
 
@@ -1476,7 +1476,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KJATQHY55X9QXRV5RT2SBGSE"
+        "tool_01KJAXYD6HBTN2PVVZWJNBHR2F"
     )
 
     display_name: Optional[str] = None
@@ -1777,7 +1777,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KJATQHY2RJCACJJ9HB3DSQK9"
+        "tool_01KJAXYD6EW0XHETE29X6EHM7Y"
     )
 
     display_name: Optional[str] = None
@@ -1972,7 +1972,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KJATQHY0R09H36MHJFB8AB8Q"
+        "tool_01KJAXYD6CFAR4RBSK144EJVJX"
     )
 
     display_name: Optional[str] = None
@@ -2171,7 +2171,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KJATQHXXWFSBWJ6ZB28WCTZ7"
+        "tool_01KJAXYD6AHQRT4QGAEEPHVS63"
     )
 
     display_name: Optional[str] = None