+++ packages/orq-rc/RELEASES.md
@@ -6139,3 +6139,13 @@ Based on:
 - [python v4.5.0-rc.8] packages/orq-rc
 ### Releases
 - [PyPI v4.5.0-rc.8] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.8 - packages/orq-rc
+
+## 2026-02-25 17:38:36
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.729.0 (2.841.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.9] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.9] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.9 - packages/orq-rc